+++ mg/go.mod
@@ -1,3 +1,3 @@
-module bal.go
+module solMod
 
 go 1.23.4
