+++ mg/utils/utils.go
@@ -1,0 +1,7 @@
+package utils
+
+import "fmt"
+
+func MyDummy() {
+	fmt.Println("dummy")
+}
